# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "所持ゲーム管理システムのバックエンドAPI仕様書",
        "title": "Yutagame Backend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/app/announcements/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public Announcements"
                ],
                "summary": "公開中のお知らせ詳細取得",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "お知らせID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "匿名訪問者ID",
                        "name": "visitorId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Announcement"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/app/banners/{id}/click": {
            "post": {
                "tags": [
                    "Public Banners"
                ],
                "summary": "バナークリック記録",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "バナーID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/app/banners/{placement}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public Banners"
                ],
                "summary": "表示位置別バナー取得",
                "parameters": [
                    {
                        "type": "string",
                        "description": "表示位置",
                        "name": "placement",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "匿名訪問者ID",
                        "name": "visitorId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Banner"
                            }
                        }
                    }
                }
            }
        },
...
        "/app/contacts": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public Contacts"
                ],
                "summary": "お問い合わせ送信",
                "parameters": [
                    {
                        "description": "お問い合わせ内容",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app.ContactRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.ContactInquiry"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/app/features/{code}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public Features"
                ],
                "summary": "公開中の特集詳細取得",
                "parameters": [
                    {
                        "type": "string",
                        "description": "特集コード",
                        "name": "code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "匿名訪問者ID",
                        "name": "visitorId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Feature"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/app/games/{code}/favorite": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public Favorites"
                ],
                "summary": "本日の「推し」状態取得",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ゲームコード",
                        "name": "code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "匿名訪問者ID",
                        "name": "visitorId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "匿名訪問者ID",
                        "name": "X-Visitor-Id",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.GameFavoriteResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "同じ訪問者・ゲームの組み合わせは日本時間で1日1回だけ登録します。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public Favorites"
                ],
                "summary": "本日の「推し」を登録",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ゲームコード",
                        "name": "code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "匿名訪問者ID",
                        "name": "X-Visitor-Id",
                        "in": "header"
                    },
                    {
                        "description": "匿名訪問者ID",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/app.GameFavoriteRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/app.GameFavoriteResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "app.ContactRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "subject": {
                    "type": "string"
                }
            }
        },
...
        "app.GameFavoriteRequest": {
            "type": "object",
            "properties": {
                "visitorId": {
                    "type": "string"
                }
            }
        },
        "app.GameFavoriteResult": {
            "type": "object",
            "properties": {
                "alreadyVoted": {
                    "type": "boolean"
                },
                "count": {
                    "type": "integer"
                },
                "favoriteDate": {
                    "type": "string"
                },
                "gameId": {
                    "type": "integer"
                }
            }
        },
...
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "model.Announcement": {
            "type": "object",
            "properties": {
                "accessCount": {
                    "type": "integer"
                },
                "bodyHtml": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "displayOrder": {
                    "description": "DisplayOrder は公開中のお知らせ同士の表示優先度（1以上が手動指定、0は未指定）",
                    "type": "integer"
                },
                "excerpt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "publishEndAt": {
                    "type": "string"
                },
                "publishStartAt": {
                    "type": "string"
                },
                "publishedAt": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "model.Banner": {
            "type": "object",
            "properties": {
                "accessCount": {
                    "type": "integer"
                },
                "clickCount": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "displayOrder": {
                    "type": "integer"
                },
                "endsAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "imageKey": {
                    "type": "string"
                },
                "linkUrl": {
                    "type": "string"
                },
                "openInNewTab": {
                    "type": "boolean"
                },
                "placement": {
                    "type": "string"
                },
                "startsAt": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "model.ContactInquiry": {
            "type": "object",
            "properties": {
                "adminNote": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "subject": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "model.Feature": {
            "type": "object",
            "properties": {
                "accessCount": {
                    "type": "integer"
                },
                "bodyHtml": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "displayOrder": {
                    "type": "integer"
                },
                "excerpt": {
                    "type": "string"
                },
                "featureGames": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.FeatureGame"
                    }
                },
                "games": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Game"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "publishEndAt": {
                    "type": "string"
                },
                "publishStartAt": {
                    "type": "string"
                },
                "publishedAt": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "thumbnailImageKey": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "model.FeatureGame": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "displayOrder": {
                    "type": "integer"
                },
                "featureId": {
                    "type": "integer"
                },
                "game": {
                    "$ref": "#/definitions/model.Game"
                },
                "gameId": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "model.Game": {
            "type": "object",
            "properties": {
                "affiliates": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.GameAffiliate"
                    }
                },
                "catchCopy": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "genre": {
                    "$ref": "#/definitions/model.Genre"
                },
                "genreId": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "imageKey": {
                    "type": "string"
                },
                "isClear": {
                    "type": "boolean"
                },
                "isFavourite": {
                    "type": "boolean"
                },
                "isPlay": {
                    "type": "boolean"
                },
                "kana": {
                    "type": "string"
                },
                "keywords": {
                    "description": "💡 凄まじいポイント：GORMに「game_keywordsという中間テーブルを使って、紐づくKeywordsを全自動でガッチャンコしてね」と1行で命令します",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Keyword"
                    }
                },
                "listPrice": {
                    "type": "integer"
                },
                "machine": {
                    "$ref": "#/definitions/model.Machine"
                },
                "machineId": {
                    "type": "integer"
                },
                "manufacturer": {
                    "description": "💡 依存関係の解消：GORMの1対多（Belongs To）リレーション群",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Manufacturer"
                        }
                    ]
                },
                "manufacturerId": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "officialSiteUrl": {
                    "type": "string"
                },
                "overview": {
                    "type": "string"
                },
                "previousRank": {
                    "type": "integer"
                },
                "rank": {
                    "type": "integer"
                },
                "rankingCount": {
                    "type": "integer"
                },
                "releaseDate": {
                    "type": "string"
                },
                "subCatch": {
                    "type": "string"
                },
                "subGenre": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "youtubeUrl": {
                    "type": "string"
                }
            }
        },
        "model.GameAffiliate": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "gameId": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "model.Genre": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "imageKey": {
                    "type": "string"
                },
                "kana": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "overview": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "model.Keyword": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "kana": {
                    "type": "string"
                },
                "keywordType": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "overview": {
                    "type": "string"
                },
                "sortOrder": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "model.Machine": {
            "type": "object",
            "properties": {
                "abbreviation": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "imageKey": {
                    "type": "string"
                },
                "kana": {
                    "type": "string"
                },
                "machineType": {
                    "type": "string"
                },
                "manufacturer": {
                    "description": "💡 依存関係の解消：GORMに「メーカー情報も連動してね」と教えるリレーション定義",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Manufacturer"
                        }
                    ]
                },
                "manufacturerId": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "overview": {
                    "type": "string"
                },
                "releaseDate": {
                    "type": "string"
                },
                "sortOrder": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "model.Manufacturer": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "imageKey": {
                    "type": "string"
                },
                "kana": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "overview": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "\"Bearer {token}\" の形式でJWTトークンを入力してください。",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}