# interface/handler/app/banner_handler.go
package app

import (
	"net/http"
	"strconv"
	"time"
	usecaseApp "yutagame-backend/application/usecase/app"
	"yutagame-backend/domain/model"
	"yutagame-backend/infrastructure/database"
	"yutagame-backend/interface/handler"

	"github.com/labstack/echo/v4"
)

type BannerHandler struct {
	bannerRepo          *database.BannerRepository
	analyticsLogUseCase *usecaseApp.AnalyticsLogUseCase
}

func NewBannerHandler(bannerRepo *database.BannerRepository, analyticsLogUseCase *usecaseApp.AnalyticsLogUseCase) *BannerHandler {
	return &BannerHandler{bannerRepo: bannerRepo, analyticsLogUseCase: analyticsLogUseCase}
}

// GetByPlacement godoc
// @Summary 表示位置別バナー取得
// @Tags Public Banners
// @Produce json
// @Param placement path string true "表示位置"
// @Param visitorId query string false "匿名訪問者ID"
// @Success 200 {array} model.Banner
// @Router /app/banners/{placement} [get]
func (h *BannerHandler) GetByPlacement(c echo.Context) error {
	items, err := h.bannerRepo.FindVisibleByPlacement(
		c.Request().Context(),
		c.Param("placement"),
		time.Now(),
	)
	if err != nil {
		return handler.RespondError(c, http.StatusInternalServerError, err)
	}
	if items == nil {
		items = []model.Banner{}
	}
	if h.analyticsLogUseCase != nil {
		for _, item := range items {
			if logErr := h.analyticsLogUseCase.RecordContentAccess(c.Request().Context(), usecaseApp.ContentAccessLogInput{
				VisitorID:   resolveVisitorID(c),
				IP:          c.RealIP(),
				ContentType: "banner",
				ContentKey:  strconv.FormatInt(item.ID, 10),
			}); logErr != nil {
				c.Logger().Warnf("banner access log write failed: %v", logErr)
			}
		}
	}
	return c.JSON(http.StatusOK, items)
}

// CountClick godoc
// @Summary バナークリック記録
// @Tags Public Banners
// @Param id path int true "バナーID"
// @Success 204
// @Router /app/banners/{id}/click [post]
func (h *BannerHandler) CountClick(c echo.Context) error {
	id, _ := strconv.ParseInt(c.Param("id"), 10, 64)
	if err := h.bannerRepo.IncrementClickCount(c.Request().Context(), id); err != nil {
		return handler.RespondError(c, http.StatusInternalServerError, err)
	}
	return c.NoContent(http.StatusNoContent)
}